--- a/src/app/[lang]/blog/article/[blog_id]/comments.tsx
+++ b/src/app/[lang]/blog/article/[blog_id]/comments.tsx
@@ -22,8 +22,8 @@ export default function BlogComments(
 
     return (
         <>
-            <div id="comments-section" className='max-w-7xl mx-auto 
-                  mt-5 text-xl 
+            <div id="comments-section" className='max-w-7xl mx-auto px-4 md:px-6 lg:px-8
+                  mt-5 text-lg 
                   md:mt-18 md:text-3xl
                 '>
                 <div>
@@ -32,8 +32,8 @@ export default function BlogComments(
                     >{BlogCommentPrompts[blogCommentPromptIndex][lang]}</h3>
                 </div>
             </div>
-            <div className='max-w-7xl mx-auto 
-                      mt-5 text-xl 
+            <div className='max-w-7xl mx-auto px-2 md:px-6 lg:px-8
+                      mt-5 text-lg 
                       md:mt-18 md:text-3xl
                     '>
                 <Comments lang={lang} message_id={message_id} singlePageMode={true} />