--- a/src/app/api/web-vitals/route.ts
+++ b/src/app/api/web-vitals/route.ts
@@ -1,0 +1,34 @@
+import { NextRequest, NextResponse } from 'next/server';
+
+export async function POST(request: NextRequest) {
+  try {
+    const webVitalData = await request.json();
+    
+    // 在这里可以将数据保存到数据库或发送到分析服务
+    console.log('📊 Web Vitals data received:', webVitalData);
+    
+    // 示例：保存到日志或数据库
+    // await saveWebVitalsToDB(webVitalData);
+    
+    // 示例：发送到外部分析服务
+    // await sendToAnalyticsService(webVitalData);
+    
+    return NextResponse.json({ success: true });
+  } catch (error) {
+    console.error('Error processing web vitals data:', error);
+    return NextResponse.json(
+      { error: 'Failed to process web vitals data' },
+      { status: 500 }
+    );
+  }
+}
+
+// 示例：保存到数据库的函数
+// async function saveWebVitalsToDB(data: any) {
+//   // 实现数据库保存逻辑
+// }
+
+// 示例：发送到外部分析服务的函数  
+// async function sendToAnalyticsService(data: any) {
+//   // 实现外部服务发送逻辑
+// }