--- a/src/app/globals.css
+++ b/src/app/globals.css
@@ -96,6 +96,25 @@
   .text-shadow-lg {
     text-shadow: 0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05);
   }
+  
+  /* Touch feedback utility classes */
+  .touch-feedback {
+    transition: transform 0.1s ease-out, background-color 0.1s ease-out;
+  }
+  
+  .touch-feedback:active {
+    transform: scale(0.95);
+  }
+  
+  /* Smooth navigation transitions */
+  .nav-item {
+    transition: all 0.2s cubic-bezier(0.4, 0, 0.2, 1);
+  }
+  
+  .nav-item:active {
+    transform: scale(0.95);
+    background-color: rgba(255, 255, 255, 0.2);
+  }
 
   body {
     font-family: 'Inter variant0', sans-serif, Arial, Helvetica, sans-serif;
@@ -119,6 +138,33 @@
     /* Prevent scrolling past footer */
     body {
       overscroll-behavior-y: none;
+    }
+  }
+  
+  /* Enhanced touch feedback for mobile devices */
+  @media (hover: none) and (pointer: coarse) {
+    button, a, .cursor-pointer {
+      -webkit-tap-highlight-color: rgba(255, 255, 255, 0.1);
+      tap-highlight-color: rgba(255, 255, 255, 0.1);
+    }
+    
+    /* Add visual feedback for active state */
+    button:active, a:active, .cursor-pointer:active {
+      transform: scale(0.95);
+      transition: transform 0.1s ease-out;
+    }
+    
+    /* Ensure smooth transitions on touch devices */
+    * {
+      -webkit-touch-callout: none;
+      -webkit-user-select: none;
+      user-select: none;
+    }
+    
+    /* Allow text selection in content areas */
+    p, span, div[contenteditable], input, textarea {
+      -webkit-user-select: text;
+      user-select: text;
     }
   }