--- a/src/types/gtag.d.ts
+++ b/src/types/gtag.d.ts
@@ -1,0 +1,13 @@
+// Google Analytics 4 (gtag) type declarations
+
+declare global {
+  interface Window {
+    gtag: (
+      command: 'config' | 'event' | 'set' | 'get',
+      targetId: string,
+      config?: Record<string, string | number | boolean | undefined>
+    ) => void;
+  }
+}
+
+export {};