… docs/WEB_VITALS_USAGE.md
… package.json
… src/components/Analytics/GoogleAnalytics.tsx
… src/components/Analytics/ReadingTimeTracker.tsx
… src/components/Comments/Comments.tsx
… src/components/Comments/index.tsx
… src/components/Common/Foot.tsx
… src/components/Common/Nav.tsx
… src/components/util/OptimizedImage.tsx
… src/components/util/WebVitalsDashboard.tsx
… src/components/util/WebVitalsProvider.tsx
… src/hooks/usePerformance.ts
… src/lib/gtag.ts
… src/lib/performance.ts
… src/lib/web-vitals.ts