## Summary

This PR implements a comprehensive Google Analytics 4 and performance monitoring system for the blog, along with significant mobile UX improvements and code maintainability enhancements.

### 🎯 Key Features

#### Google Analytics 4 Integration
- **Tracking ID**: G-N7P75W40JD configured for production
- **Event Tracking**: Comprehensive user interaction tracking across all components
- **Performance Metrics**: Web Vitals integration with real-time monitoring
- **Privacy Compliant**: IP anonymization and GDPR-friendly settings

#### Advanced Analytics Features
- **Reading Time Tracking**: Article engagement metrics with milestone tracking (15s, 30s, 60s, 120s)
- **Friend Link Analytics**: Detailed tracking of external link clicks with categorization
- **Scroll Depth Monitoring**: Track user engagement at 25%, 50%, 75%, 90% scroll depths
- **Error Tracking**: Automatic JavaScript and Promise error monitoring
- **User Journey Analytics**: Page visibility changes and time-on-page tracking

#### Performance Monitoring System
- **Web Vitals Dashboard**: Real-time LCP, INP, CLS, FCP, TTFB monitoring
- **Performance Hooks**: Custom React hooks for component-level performance tracking
- **Bundle Optimization**: Code splitting and webpack optimizations
- **Image Optimization**: WebP/AVIF support with lazy loading

#### Mobile UX Improvements
- **Touch Feedback**: Enhanced mobile navigation with visual touch responses
- **Responsive Layouts**: Mobile-first design improvements for Comments and Friends pages
- **Optimized Spacing**: Better content-to-screen ratio on mobile devices
- **Performance**: Reduced bundle size and improved loading times

#### Code Quality & Maintainability
- **Comment Translation**: All Chinese comments translated to English
- **TypeScript Safety**: Comprehensive type definitions for GA4 integration
- **React Best Practices**: Optimized component rendering with useCallback, useMemo, memo
- **ESLint Compliance**: All linting rules satisfied

### 📊 Analytics Coverage

#### Tracked Events
- Navigation clicks and external links
- Blog article interactions and reading patterns
- Comment submissions and views
- Friend link clicks by category
- Language switching patterns
- Search and filter usage
- Error occurrences and performance issues

#### Performance Metrics
- **Core Web Vitals**: All metrics tracked and reported to GA4
- **Custom Metrics**: Component render times and user interaction latency
- **Business Metrics**: Content engagement and user retention indicators

### 🚀 Technical Implementation

#### New Components
- `GoogleAnalytics`: GA4 script integration with enhanced tracking
- `PageTracker`: Automatic page view and route change tracking
- `ReadingTimeTracker`: Article engagement time monitoring
- `WebVitalsProvider` & `WebVitalsDashboard`: Performance monitoring UI
- `OptimizedImage`: Lazy loading with WebP/AVIF support

#### Enhanced Components
- `Comments`: Mobile layout optimization and GA tracking
- `Nav`: Touch feedback and comprehensive interaction tracking
- `Pagination`: Click tracking and mobile improvements
- `Foot`: External link tracking for social media and RSS
- `Friends`: Detailed friend link analytics with categorization

### 🔧 Build & Performance

#### Bundle Optimization
- **Code Splitting**: Dynamic imports for Analytics components
- **Tree Shaking**: Optimized webpack configuration
- **Compression**: Enhanced build output with better caching

#### Performance Results
```
Route (app)                    Size    First Load JS
├ ƒ /[lang]                   9.69 kB    347 kB
├ ƒ /[lang]/friends           4.62 kB    350 kB
├ ƒ /[lang]/blog/[blog_page]  6.72 kB    344 kB
+ First Load JS shared by all  332 kB
```

### 🧪 Testing & Quality Assurance

- ✅ TypeScript compilation: No errors
- ✅ ESLint checks: All rules satisfied
- ✅ Build process: Successful production build
- ✅ Performance: Web Vitals within acceptable ranges
- ✅ Mobile responsiveness: Tested across devices

### 📈 Expected Impact

#### User Experience
- Improved mobile navigation and interaction feedback
- Faster page loads through optimized images and code splitting
- Better content accessibility with responsive design

#### Analytics Insights
- Comprehensive user behavior analysis
- Performance bottleneck identification
- Content engagement optimization opportunities
- Mobile vs desktop usage patterns

#### Development Workflow
- Better code maintainability with English comments
- Type-safe GA4 integration
- Performance monitoring alerts
- Error tracking and debugging capabilities

🤖 Generated with [Claude Code](https://claude.ai/code)